Fixes SSL websocket disconnects with client
Websocket was configured with a socket timeout of 10 seconds which was
causing the SSL connection to close with timeout errors. This is due to
an issue with the websocket-client library not catching an SSL timeout
error properly.  This patch adds a check for SSL time out in the event
it is not caught and presented as a WebSocketTimeoutException.

Under normal circumstances the websocket will read and issue a timeout
if no data was received on the socket over the duration of the timeout.
This happens quite often in the thread because ODL is only sends port
status updates on an event.  The thread should just continue to poll in
the event of a time out (no data received).

Closes-Bug: 1744759

Change-Id: I01f81befdd19d8ad914761c7013f10f2b5d69602
Signed-off-by: Tim Rozet <trozet@redhat.com>