Fix missing parentid on rule delete journal record
This fixes an error raised when deleting a security group if there are
pending rules to be deleted of that group as records in the journal
that were not triggered by the group deletion operation itself so that
they did not have the parent security group id as data of the record.

This happens when the rule and the group are deleted independently but
in rapid succession. It does not happen when the rules are deleted as a
consequence of deleting the group since the parentid is correctly set
in that case, neither when the group is deleted some time after the
rules when their records might have already been processed and removed.

This problem was encountered in OPNFV functest environment but may also
fix https://jira.opendaylight.org/browse/NETVIRT-896 and complement
change #533706.

Change-Id: I5ccdbdea9d803334c8ae9416b40e3c4b48a8144b
Closes-Bug: 1744966
Signed-off-by: Jaime Caamaño Ruiz <jcaamano@suse.com>