Wait for worker start before testing in JournalPeriodicProcessorTest
These tests are susceptible to a race condition:
  - JournalPeriodicProcessorTest.test_handle_sighup_gracefully
  - JournalPeriodicProcessorTest.test_reset_called_on_sighup

The race condition happens when the worker is SIGHUP before both the signal
processing is registered and the worker is started. In that case there
is no handle by the IPC set by the mock and therefore the testing
process timeouts waiting for the event to happen.

This patch removes the time element from the test by making it event
based.

Change-Id: Ie5310bd026f29e2a5b61eaa1109fb051e09b3258