Move zuul_copy_output to be a job variable
The zuul_copy_output variable is designed to be able to be used as
a zuul job variable so that zuul dictionary merging will work. However,
it's currently being set in the playbook rather than as a job variable,
so it's not possible to supplement it in a child job.

Move it to be a job variable. Also remove the wrapping {} as they should
not be needed to make zuul_copy_output a dictionary.

Change-Id: I27da22bd48dbe7252db9f9d22f3087b070fb26ac
Depend-On: https://review.openstack.org/#/c/539686/