Remove vpnservice conditional from pre test hook
As tempest nitrogen and carbon jobs have been migrated to v3
pre run take care of pre test hooks which were present for all
jobs in devstack/pre_test_hook.sh. so hooks for vpn service shouldn't
be there in it. This purging and recreation of network was only for
jobs which actually deploy devstack and run tempest tests against that.
So fullstack/functiona/rally jobs won't need recreation of neutron networks.

Change-Id: I36c84888e1d24ca0522054e6967b3c7479432e7e