Switch to a blacklist file for test excludes
This commit switches to use a blacklist file to exclude tests. It's much
more explicit and less error prone than the giant regex used before.
Each line in the blacklist file is treated as a separate regex, and
after the initial set of tests to run is found (from any cli regex or
whitelist file regexes) tests that match an re.search on each line are
removed from that set.

Change-Id: I83ec00dd80ada20b1cf7840d2110c68cba3d3efa
Depends-On: https://review.openstack.org/547606
Co-Authored-By: Michel Peterson <michel@redhat.com>