Allow Tempest CI to add txt extensions to ini files
When looking at logs of the CI it's useful to be able to read certain
files in the browser. Most plain text files get a txt extension to allow
easier consumption. INI files lack this and thus are downloaded.

This patch allows the Tempest CI to add this extension to INI files.

Change-Id: I05ce7e28148d4edaf32049cc20065a5de5f5dbfc