Abstract how retriable methods are tested
Until now retriable methods where expected to start a transaction but
were asserted by looking for a call to `Session.begin`. This was a good
interim solution, however it is implementation dependant and it does not
actually test for behavior. Also, moving into the new DB pattern,
i.e. enginefacade, where transactions are handled differently, this is a
must.

This patch modifies the test so it actually test that the transaction is
reverted on failure, which is what is needed for a retriable method,
i.e. idempotence is expected.

Change-Id: I0d11ec456d36852a5a906f1be83e139a9f403c00