Update DB functions to handle contexts
With the new enginefacade [1] all functions that interact with the DB
are expected to recieve a context if they are to do operations inside a
transaction. It is possible to have a mix of session and context aware
functions but it's better to standarize to only context aware functions.
As we migrate into the new enginefacade pattern, this new pattern should
become the norm.

This patch modifies DB functions to handle only context and it is a
preliminary patch to migrate to the new enginefacade pattern.

[1]:
http://specs.openstack.org/openstack/oslo-specs/specs/kilo/make-enginefacade-a-facade.html

Change-Id: I681a02010908373c823d0aecba81c1906da2ec5a