Remove incorrect DB retry decorators
The DHCP driver has a retry decorator that expects the function to not
receive a context -yet it does- and that does not create a transaction
or savepoint to be able to retry. These two behaviors are inherently
incorrect.

This patch removes the retry decorator. This is done since it wasn't
doing anything, because of the previously explained reasons. And because
it standarizes to what other drivers are currently doing. The retry can
be added later if we found it's needed.

Change-Id: I8c2f43399913fb6066069dde8017ecffdee7c9f0