Fix tox installation of neutron
There has been a change to put neutron on the upper constraints file [1]
This change breaks the way we consume it, now there is no longer a need
to use `tox_install.sh` for neutron and it can be consumed directly with
pip install -since it is now published in PyPI, it can be consumed from
there.

This patch puts neutron on the requirements file and removes it from the
`tox_install.sh` script.

[1] https://review.openstack.org/550475/

Change-Id: Iab198fa31c6f0a895b151ce8cbe819da8cc16e78