Fix CI issues that block the gate
This patch is the composition of two different patches:

1. Fix PEP8 pylint errors
=========================

The linter started complaining about a relative-import issue with the
library six. Once that import was fixed it started complaining on a
particular function of the module imported with a
too-many-function-args.

This patch fixes the relative-import and disables the
too-many-function-args locally for that function, as it is a problem
with the library and not a real error.

2. Fix functional migration tests
=================================

The functional migration tests has ignore for tables that are external
and not related to the current repo and now the migration tests started
failing on tables that weren not present in that list. Also, the failing
migrations are indeed external but are also related to the repo, so
their categorization is incorrect.

This patch separates between external that are not related to the repo
and ignored tables that are related but are not managed by us. In
addition, it adds the two missing matches so it does not fail on them.

Change-Id: Ieaad6b61f55f86f54ecc27aae74c3b2495ec6eb5