Remove attribute db_session from ODLBaseDbTestCase
With enginefacade DB sessions are attached to a context. This is handled
by the different helper enginefacade functions in a way that session and
transaction creation happen transparent to the consumer. The fact that
we were creating a copy of the session on `self.db_session` means that
when enginefacade abstracts the session management we might end with two
different sessions in different places of the code. The correct way to
do it is access the context's session when needed directly from the
context. That way we ensure we are using the proper one.

This patch removes the attribute and makes modifications to depend in
the context's session and be future proof for the next patches that will
introduce enginefade.

Change-Id: I26034dd6003effcfc51d431a1c22e98c5109486c