Remove mock for odl_features.init()
It is currently mocked because by the justification that calling it
would try to fetch features from ODL. However, this is not needed since
there is a configuration key that can be used to override the fetch
behavior and instead force which features are enabled.

This patch removes the unnecesary and counter-intuitive mock, and it
uses the configuration key instead.

Change-Id: I0b03fc3708c2026afa066ef1836ada5cc94b672d