Retry journal recording in L3
Since journal recording can fail, and is not run in the same
transaction as router/FIP operations, it will lead to the resource
(router or FIP) to remain altered while a journmal record hasn't been
created.

L3 Flavors should solve this issue comprehensively, but we need a
solution until it's stable, which can also be backported to stable
branches. Hence, this fix should address 99% or the cases.

Change-Id: Ibef0db9a58d86a85ffd948a35db983863999d319
Closes-Bug: #1738246