Fixes for the CI
This is a combination of two commits:

Commit #1: Add missing devstack function

The change [1] added an expectation of the function `conductor_conf` to
exist. In the cases of functional/fullstack testing this function
does not exist, thus breaking the devstack flow.

This patch adds a no-op function to functional/fullstack testing to
avoid this issue.

[1]: https://review.openstack.org/#/c/576249/

Commit #2: Switch to stestr

According to Openstack summit session [2] stestr is
maintained project to which all Openstack projects
should migrate.

[2] https://etherpad.openstack.org/p/YVR-python-pti

Change-Id: Ic5622ce52b7ed595c7bd0a1b9fcbaa1c6e9f781d
Co-Authored-By: shubhendu <shubhendu.poothia@nectechnologies.in>

Change-Id: I3f91833c778fcee373dcdfb89ae9e9bbf3dd7a59