Assure that PeriodicTask passes context as args
Enginefacade expects that the first argument of functions are either
self, cls or the context. At the same time, PeriodicTask was sending
context as a kwargs called context. Because the introspection that
enginefacade does only happens on args this was causing an error with
PeriodicTask.

This patch modifies the behavior of PeriodicTask so it only sends the
context as enginefacade expects it -an args.

Change-Id: I7edb4c3284a358fba80e4fe68a0cd0f08f809b1a
Closes-Bug: #1781787