Add docs requirements to venv
Prompted by a change from OpenStack TC the doc requirements live
separately. When that change entered this repo it slipped to add that
requirements to the venv to be able to do docs related operations (e.g.
create release notes). Thus currently the documented usage [1] does not
work for n-odl.

This patch adds the docs requirements to venv, restoring that
functionality.

[1]: https://docs.openstack.org/reno/latest/user/usage.html

Change-Id: I970d27c80343f185f7b2ec92da651f1fb6dcc986