Consume DB retry decorator from neutron-lib
In a process of moving common things out of neutron and into neutron-lib
there are patches being issued to gradually move them. The last patch
that dealt with retry decorators [1] was incomplete, but somehow the
gate succeeded and it still succeeds, although it fails locally.

This patch fixes this problem by consuming the retry decorator from
neutron-lib as it was originally intended. In addition, it bumps the
requirements to the version required for this to work.

[1]: https://review.openstack.org/570028

Change-Id: I73f39190cac5f17412ff272a1bdfecfa7d7c3f68