Remove dns-integration from NETWORK_API_EXTENSIONS
As of now DNS-integration is not supported by networking-odl,
which cause failure of tempest test and require_ext decorator is
not very smart enough to figure out if extension is actually
enabled or not, it just reads the tempest conf and decides regardless
of actual environment. This patch removes dns-integration from
tempest-conf.

Change-Id: If600cc0989ba16606291e83bcc029537d57e5017
Related-Bug: #1785656