Removing the fullstack tests
As discussed on the IRC meeting [1], these tests have no value and are
100% failing all the time now.
Additionally in the stadium requirements [2] there's no requirement to
have these tests.

It makes no sense to keep them as such.

Additionally, networking_odl/tests/contrib directory is removed.
At first glance it contained mentions of fullstack which needed updated,
but upon further research it was found that the files in it weren't used
so it's removed altogether.

[1] https://meetings.opendaylight.org/opendaylight-neutron/2018/weekly_meeting/opendaylight-neutron-weekly_meeting.2018-07-30-15.03.txt
[2] https://docs.openstack.org/neutron/latest/contributor/stadium/governance.html#when-is-a-project-considered-part-of-the-stadium

Change-Id: I0c78abb3157a7813b87de168bfb39cab2e56d725