Dont disable services that are not enabled
Both neutron lib functions is_neutron_enabled and
is_neutron_legacy_enabled check for disabled services. If a particular
service is disabled, it will be assumed that the component that service
is part of is enabled. For example, if any 'neutron-*' service is
disabled, devstack will assume that neutron (non-legacy) is enabled.

This patch changes to disable services only if they wrere previously
enabled to avoid any assumption based on disabled services that were
never enabled on the first place.

Change-Id: If29ccbc5a7d80ad046ece26d822a6f55aa73f0b7