Remove test-requirements from functional tox env
Functional tests are meant to be closer to a deployment than jobs that are
doing unit testing or the like. As such it needs to have the minimal amount
of testing related packages that it needs. As it is, test-requirements should
not be included. 

This patch makes use of the tests/functional/requirements.txt and removes
test-requirements.txt. 

A good use case of this is to test a deployment without ceilometer, as it would
be deployed if ceilometer will become an extra dependency. The solution doesn't
come without drawbacks. An example is that ceilometer code will not get tested
functionally. However, this is acceptable since the community agrees that 
ceilometer driver should probably be removed from n-odl. If that decision were
to change it's always possible to create a separate set of jobs that do include
ceilometer and test with ceilometer as well.

Change-Id: I9a52f34d5fa0bb4066ecb51792f5e291c441ad01
Signed-off-by: Manuel Buil <mbuil@suse.com>