Fix gate failure
When creating a tox env, networking-bgpvpn train version is installed.
It has a dependency of horizon<17.0.0 (train) which requires Django<2.1
and it is not compatible with u-c and horizon ussuri. It leads to a tox
creation failure. We need to install ussuri version of networking-bgpvpn.
A beta1 of networking-bgpvpn has been released, so we can use it.

neon release of ODL is no more available as snapshot, but as a released
so the neon functional job was changed to use latest release instead of
a snapshot.

This patch is based on https://review.opendev.org/708829 from amotoki,
again thanks to him.

Change-Id: I0a798e5ab92ad079e070c577360147ab8980bf0a