Update pylint version
Gate started to fail due to typed-ast was constrained in
upper-constraints.txt [1]. typed-ast is pulled in by astroid, which is
pulled in by pylint.

To have the new version of upper-constrained typed-ast (1.5.1) at least
astroid 2.8.5 needs to be installed. Pylint pins astroid version
explicitly, so the first pylint vesion that satisfies astroid >= 2.8.5
is pylint 2.12.0 (which pins astroid 2.9.0 directly).

To untangle this situation pylint is updated to 2.12.0 version and
astroid is removed from test-requirements.txt as it is pinned within
pylint anyway.

mccabe also needed an update in lower-constraints.txt because
pylint 2.12.0 depends on mccabe<0.7 and >=0.6

Due to pylint bump the following rules were added to the disable list in
.pylintrc:
* consider-using-f-string
* super-with-arguments
and the following rules were aplied to the code:
* inconsistent-return-statements
* raise-missing-from
* arguments-renamed
* use-dict-literal
* unspecified-encoding
* deprecated-decorator

[1] I1734720c04e84d3fbb73b5564cd21c80ca85b334

Change-Id: Ic8c2583b553eae360fe4e982be0a7b261f6df91f