Fix tox4 issues and define project upper constraints
With tox4 allowlist_externals is more strictly checked, so
fix it where necessary and fix pylint version.

This patch also adds the needed versions for the development environment
projects, specifying the maximum version supported in this release.

This patch also removes the lower-constraints job, and sets the
required-projects for py39 job to point to stable/yoga branches.

The grenade, functional and tempest jobs (broken for a long time), are
also removed from the CI queues and their definitions.

Related-Bug: #1999558
Change-Id: Ie795988c0a3528397e7b76aee2cafa440625bd4e