Fix tox4 issues and define project upper constraints
With tox4 allowlist_externals is more strictly checked, so
fix it where necessary and fix pylint version.

This patch also adds the needed versions for the development environment
projects, specifying the maximum version supported in this release.

The grenade, functional and tempest jobs (broken for a long time), are
also removed from the CI queues and their definitions.

Related-Bug: #1999558
Change-Id: I3409272ce0fb9e1033f1621ac1d7db0a2b5d73c3